refactor : 상대경로에서 절대경로로 변경
