--- src/pages/DetailPage/useSelectedPost.ts
@@ -1,5 +1,0 @@
-import { useSelector } from 'react-redux';
-import { RootState } from '@/store';
-
-export const useSelectedPost = () =>
-  useSelector((state: RootState) => state.postDetail.post);
